Rename TimeStamp->Status
